+++ types/doc_info.go
@@ -28,8 +28,8 @@ type DocInfosShard struct {
 
 // DocInfo document info
 type DocInfo struct {
-	Fields       interface{}
-	TokenLengths float32
+	Fields    interface{}
+	TokenLens float32
 }
 
 /// inverted_index.go
@@ -37,7 +37,7 @@ type DocInfo struct {
 // InvertedIndexShard 反向索引表([关键词]反向索引表)
 type InvertedIndexShard struct {
 	InvertedIndex    map[string]*KeywordIndices
-	TotalTokenLength float32 //总关键词数
+	TotalTokenLen float32 //总关键词数
 	sync.RWMutex
 }
 
@@ -46,5 +46,5 @@ type KeywordIndices struct {
 	// 下面的切片是否为空，取决于初始化时IndexType的值
 	DocIds      []uint64  // 全部类型都有
 	Frequencies []float32 // IndexType == FrequenciesIndex
-	Locations   [][]int   // IndexType == LocationsIndex
+	Locations   [][]int   // IndexType == LocsIndex
 }